Fix catch button release too early bug.
I had noticed that it was very hard to press the catch button at the exact
time the last ball LED was illuminated, so I allowed a leeway and a valid
catch can be made if the ball one of the last 2 LEDs (see the ENDZONE
constant).

While playing the game I noticed that if you pressed and released the catch
button exactly when the 2nd last ball LED was illuminated but that by the time
the last ball LED was illuminated you had released the button, you would miss
the catch and the other player would get a point, which isn't correct.

The BALLSAVE variable was being reset on every pass through the main loop,
in other words, on every illumination of each ball LED. This fix modifies the
code such that the BALLSAVE variable is only reset when the direction of play
changes, in other words, the BALLSAVE variable is not reset during one side's
entire play.